Changed the all endpoint to index
